+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 ### Enhancements
 - Adds improved constructors for `SuperwallOptions`,`PaywallOptions` and `PaywallPresentationHandler`, allowing a DSL like usage,i.e. ` PaywallPresentationHandler { onPresent { ... } }` 
 
+### Fixes
+- Ensures poster is always visible on video preview
+- Fix bug with loading not dismissing post purchase if user remains in paywall
+- Fix warning about old edge-to-edge API usages
+- Fix shimmerview theme warnings and memory leak
+
 ## 2.7.0
 
 ### Enhancements